+++ io-service/src/main/java/com/example/ioservice/dto/WishlistItemDTOs/WishlistItemAddDTO.java
@@ -1,0 +1,6 @@
+package com.example.ioservice.dto.WishlistItemDTOs;
+
+public record WishlistItemAddDTO(Long wishlistId,
+                                 Long userId,
+                                 Long productId) {
+}