+++ io-service/src/main/java/com/example/ioservice/dto/WishlistItemDTOs/WishlistItemDTO.java
@@ -1,0 +1,9 @@
+package com.example.ioservice.dto.WishlistItemDTOs;
+
+import java.time.LocalDateTime;
+
+public record WishlistItemDTO(Long wishlistItemId,
+                              Long wishlistId,
+                              Long productId,
+                              LocalDateTime createdAt) {
+}