+++ io-service/src/main/java/com/example/ioservice/repository/UserRepository.java
@@ -3,6 +3,8 @@ package com.example.ioservice.repository;
 import com.example.ioservice.model.UserModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.util.Optional;
+
 public interface UserRepository extends JpaRepository<UserModel, Long> {
     UserModel findByUsername(String username);
 }