+++ market-service/src/main/java/com/example/marketservice/controller/MarketController.java
@@ -16,11 +16,16 @@ import java.util.List;
 public class MarketController {
     private final MarketService marketService;
 
+    @GetMapping("/test_market")
+    @ResponseStatus(HttpStatus.OK)
+    public ResponseEntity<String> startCheck() {
+        return ResponseEntity.ok("Market service working");
+    }
+
     @PostMapping("/add_product")
     public ResponseEntity<ProductDto> addProduct(@RequestBody ProductDto productDto) {
         return marketService.addProduct(productDto);
     }
-
 
     @GetMapping("/see_all_products")
     List<ProductDto> seeAllProducts() {
@@ -37,10 +42,64 @@ public class MarketController {
         return marketService.searchByName(name);
     }
 
-    @GetMapping("/test_market")
-    @ResponseStatus(HttpStatus.OK)
-    public ResponseEntity<String> startCheck() {
-        return ResponseEntity.ok("Market service working");
+    @PostMapping("/wishlist/add")
+    public ResponseEntity<?> addWishlist(@RequestBody WishlistAddDTO wishlistAddDTO) {
+        return marketService.addWishlist(wishlistAddDTO);
+    }
+
+    @GetMapping("/wishlist/{id}")
+    public ResponseEntity<?> getWishlistById(@PathVariable("id") Long id, @RequestParam("userId") Long userId) {
+        return marketService.getWishlistById(id, userId);
+    }
+
+    @PostMapping("/wishlist/delete/{id}")
+    public ResponseEntity<?> deleteWishlist(@PathVariable("id") Long id, @RequestParam("userId") Long userId) {
+        return marketService.deleteWishlist(id, userId);
+    }
+
+    @PostMapping("/wishlist-item/add")
+    public ResponseEntity<?> addWishlistItem(@RequestBody WishlistItemAddDTO wishlistItemAddDTO) {
+        return marketService.addWishlistItem(wishlistItemAddDTO);
+    }
+
+    @GetMapping("/wishlist-item/{id}")
+    public ResponseEntity<?> getWishlistItemById(@PathVariable("id") Long id, @RequestParam("userId") Long userId) {
+        return marketService.getWishlistItemById(id, userId);
+    }
+
+    @GetMapping("/wishlist/name/{name}")
+    public ResponseEntity<?> getWishlistItems(@PathVariable("name") String name, @RequestParam("userId") Long userId) {
+        return marketService.getWishlistItems(name, userId);
+    }
+
+    @PostMapping("/wishlist-item/delete/{id}")
+    public ResponseEntity<?> deleteWishlistItem(@PathVariable("id") Long id, @RequestParam("userId") Long userId) {
+        return marketService.deleteWishlistItem(id, userId);
+    }
+
+    @PostMapping("/review/add")
+    public ResponseEntity<?> addReview(@RequestBody ReviewAddDTO reviewAddDTO) {
+        return marketService.addReview(reviewAddDTO);
+    }
+
+    @GetMapping("/review/{id}")
+    public ResponseEntity<?> getReviewById(@PathVariable("id") Long id) {
+        return marketService.getReviewById(id);
+    }
+
+    @GetMapping("/review/user/{id}")
+    public ResponseEntity<?> getReviewsByUserId(@PathVariable("id") Long id) {
+        return marketService.getReviewsByUserId(id);
+    }
+
+    @GetMapping("/review/product/{id}")
+    public ResponseEntity<?> getReviewsByProductId(@PathVariable("id") Long id) {
+        return marketService.getReviewsByProductId(id);
+    }
+
+    @PostMapping("/review/delete/{id}")
+    public ResponseEntity<String> deleteReview(@PathVariable("id") Long id, @RequestParam("userId") Long userId) {
+        return marketService.deleteReview(id, userId);
     }
 
     @PostMapping("/add_product_to_cart")