+++ market-service/src/main/java/com/example/marketservice/dto/ReviewDTOs/ReviewAddDTO.java
@@ -1,0 +1,10 @@
+package com.example.marketservice.dto.ReviewDTOs;
+
+import com.example.marketservice.enums.ReviewRatingsEnum;
+
+public record ReviewAddDTO(Long userId,
+                           Long productId,
+                           String title,
+                           String description,
+                           ReviewRatingsEnum reviewRatingsEnum) {
+}