+++ market-service/src/main/java/com/example/marketservice/dto/WishlistDTOs/WishlistAddDTO.java
@@ -1,0 +1,5 @@
+package com.example.marketservice.dto.WishlistDTOs;
+
+public record WishlistAddDTO(String wishlistName,
+                             Long userId) {
+}