+++ market-service/src/main/java/com/example/marketservice/dto/WishlistDTOs/WishlistDTO.java
@@ -1,0 +1,6 @@
+package com.example.marketservice.dto.WishlistDTOs;
+
+public record WishlistDTO(Long wishlistId,
+                          String wishlistName,
+                          Long userId) {
+}