+++ market-service/src/main/java/com/example/marketservice/feignclient/IOClient.java
@@ -21,6 +21,41 @@ public interface IOClient {
     @GetMapping("/search")
     List<ProductDto> searchByName(@RequestParam("name") String name);
 
+    @PostMapping("/wishlist/add")
+    public WishlistDTO addWishlist(@RequestBody WishlistAddDTO wishlistAddDTO);
+
+    @GetMapping("/wishlist/{id}")
+    public WishlistDTO getWishlistById(@PathVariable("id") Long id, @RequestParam("userId") Long userId);
+
+    @PostMapping("/wishlist/delete/{id}")
+    public void deleteWishlist(@PathVariable("id") Long id, @RequestParam("userId") Long userId);
+
+    @PostMapping("/wishlist-item/add")
+    public WishlistItemDTO addWishlistItem(@RequestBody WishlistItemAddDTO wishlistItemAddDTO);
+
+    @GetMapping("/wishlist-item/{id}")
+    public WishlistItemDTO getWishlistItemById(@PathVariable("id") Long id, @RequestParam("userId") Long userId);
+
+    @GetMapping("/wishlist/name/{name}")
+    public List<WishlistItemDTO> getWishlistItems(@PathVariable("name") String name, @RequestParam("userId") Long userId);
+
+    @PostMapping("/wishlist-item/delete/{id}")
+    public void deleteWishlistItem(@PathVariable("id") Long id, @RequestParam("userId") Long userId);
+
+    @PostMapping("/review/add")
+    public ReviewDTO addReview(@RequestBody ReviewAddDTO reviewAddDTO);
+
+    @GetMapping("/review/{id}")
+    public ReviewDTO getReviewById(@PathVariable("id") Long id);
+
+    @GetMapping("/review/user/{id}")
+    public List<ReviewDTO> getReviewsByUserId(@PathVariable("id") Long id);
+
+    @GetMapping("/review/product/{id}")
+    public List<ReviewDTO> getReviewsByProductId(@PathVariable("id") Long id);
+
+    @PostMapping("/review/delete/{id}")
+    public void deleteReview(@PathVariable("id") Long id, @RequestParam("userId") Long userId);
     @PostMapping("/add_product_to_cart")
     String addProductToCart(@RequestBody AddProductToCartDto addProductToCartDto);